LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/query-evaluation-style-expected.txt


PASS style((--foo: bar))
PASS style((--baz: qux))
PASS style((unknown))
PASS unknown((--foo: bar))
PASS style(not (--foo: bar))
PASS style(not (--baz: qux))
PASS style(not (unknown))
PASS style((--foo: bar) and (--foo: bar))
PASS style((--foo: bar) and (--foo: bar) and (--foo: bar))
PASS style((--baz: qux) and (--baz: qux))
PASS style((--baz: qux) and (--foo: bar) and (--foo: bar))
PASS style((--foo: bar) and (--baz: qux) and (--foo: bar))
PASS style((--foo: bar) and (--foo: bar) and (--baz: qux))
PASS style((unknown) and (--foo: bar) and (--foo: bar))
PASS style((--foo: bar) and (unknown) and (--foo: bar))
PASS style((--foo: bar) and (--foo: bar) and (unknown))
PASS style((--foo: bar) or (--foo: bar))
PASS style((--foo: bar) or (--foo: bar) or (--foo: bar))
PASS style((--baz: qux) or (--baz: qux))
PASS style((--baz: qux) or (--foo: bar) or (--foo: bar))
PASS style((--foo: bar) or (--baz: qux) or (--foo: bar))
PASS style((--foo: bar) or (--foo: bar) or (--baz: qux))
PASS style((unknown) or (--foo: bar) or (--foo: bar))
PASS style((--foo: bar) or (unknown) or (--foo: bar))
PASS style((--foo: bar) or (--foo: bar) or (unknown))
PASS style((unknown) or (--baz: qux) or (--foo: bar))
PASS style(not ((--foo: bar) and (--foo: bar)))
PASS style(not ((--foo: bar) and (--baz: qux)))
PASS style((--foo: bar) and (not ((--baz: qux) or (--foo: bar))))
PASS style((--baz: qux) or (not ((--baz: qux) and (--foo: bar))))
PASS style((--baz: qux) or ((--baz: qux) and (--foo: bar)))
FAIL style(5 >= calc(3 + 1)) assert_equals: expected "true" but got "false"
FAIL style(1px <= 1em) assert_equals: expected "true" but got "false"
FAIL style(5 >= 3) assert_equals: expected "true" but got "false"
FAIL style(0 = 0px) assert_equals: expected "true" but got "false"
PASS style(0 = 0deg)
FAIL style(0 >= 0px) assert_equals: expected "true" but got "false"
FAIL style(0 < 3px) assert_equals: expected "true" but got "false"
FAIL style(3px >= 3px) assert_equals: expected "true" but got "false"
PASS style(3px > 3px)
FAIL style(1em > 1px) assert_equals: expected "true" but got "false"
PASS style(1em <= 1px)
FAIL style(3turn > 3deg) assert_equals: expected "true" but got "false"
FAIL style(3% >= 3%) assert_equals: expected "true" but got "false"
FAIL style(3s > 3ms) assert_equals: expected "true" but got "false"
FAIL style(3dppx > 96dpi) assert_equals: expected "true" but got "false"
FAIL style(3 <= --x) assert_equals: expected "true" but got "false"
FAIL style(--x >= 3) assert_equals: expected "true" but got "false"
FAIL style(--x <= --x) assert_equals: expected "true" but got "false"
FAIL style(--x >= --y) assert_equals: expected "true" but got "false"
FAIL style(--length > 3px) assert_equals: expected "true" but got "false"
FAIL style(--x > 3px) assert_equals: expected "true" but got "false"
FAIL style(--number >= 3) assert_equals: expected "true" but got "false"
FAIL style(--x >= 1) assert_equals: expected "true" but got "false"
FAIL style(--percentage > 3%) assert_equals: expected "true" but got "false"
FAIL style(--x > 3%) assert_equals: expected "true" but got "false"
FAIL style(--angle < 1turn) assert_equals: expected "true" but got "false"
FAIL style(--x < 1turn) assert_equals: expected "true" but got "false"
FAIL style(--time <= 1000ms) assert_equals: expected "true" but got "false"
FAIL style(--x <= 1000ms) assert_equals: expected "true" but got "false"
FAIL style(3dppx > --resolution) assert_equals: expected "true" but got "false"
FAIL style(3dppx > --x) assert_equals: expected "true" but got "false"
FAIL style(--x >= calc(3px + 3px)) assert_equals: expected "true" but got "false"
FAIL style(--x <= var(--x)) assert_equals: expected "true" but got "false"
FAIL style(calc(var(--x) + 1) >= var(--y)) assert_equals: expected "true" but got "false"
FAIL style(var(--x) >= --x) assert_equals: expected "true" but got "false"
FAIL style(--x <= 10em) assert_equals: expected "true" but got "false"
PASS style(--x + 1 >= --y)
PASS style(--x >= --y + 1)
PASS style(calc(--x + 1) >= --y)
PASS style(--x >= calc(--y + 1))
PASS style(3px > 3)
PASS style(3em > 3deg)
PASS style(1px >= 1%) or style(1px <= 1%)
PASS style(--length > 3)
PASS style(--number > 3px)
PASS style(--x >= 3ms)
FAIL style(--length <= 30px) assert_equals: expected "true" but got "false"
FAIL style(10px <= 10px < 11px) assert_equals: expected "true" but got "false"
FAIL style(3 < --x <= 5) assert_equals: expected "true" but got "false"
FAIL style(--x >= --y > --z) assert_equals: expected "true" but got "false"
